Revert "Use ULong instead of Long as timestamp"
This reverts commit 13b75dda247f5eb6d017d0dc7a3a9ca999f6ce9a.

+++ src/main/kotlin/me/adrigamer2950/playerlogs/logs/Log.kt
@@ -10,8 +10,8 @@ import org.bukkit.entity.Player
 /**
  * Represents a log entry.
  */
-abstract class Log(val message: String, player: OfflinePlayer, val timestamp: ULong) {
-    constructor(message: String, player: OfflinePlayer) : this(message, player, Timestamp.from(Instant.now()).time.toULong())
+abstract class Log(val message: String, player: OfflinePlayer, val timestamp: Long) {
+    constructor(message: String, player: OfflinePlayer) : this(message, player, Timestamp.from(Instant.now()).time)
 
     val playerUUID = player.uniqueId // Player's UUID
 }
